Add notifyReplicationDone packet.
This packet replace notifyPartitionChange when used by storage's replicator at the end of a single partition replication.
In those case, only one partition was involved and the node concerned must match with the sender identify, causing many useless on the master side.
This packet contains only the partition number that is replicated and implies that the node's cell is now up to date.
It should be sent only by the storage node from replicator and received by master node.

git-svn-id: https://svn.erp5.org/repos/neo/trunk@1467 71dcc9de-d417-0410-9af5-da40c76e7ee4