Avoid handling events for a file descriptor more than once per _poll call.
A single fd might be reported more than once per epoll call. This causes
the associated connection to be registered multiple times, which in turn
causes problems when the first connection registration gets emptied:
handling events of the second registration will raise an error when trying
to fetch a packet from an empty list.

git-svn-id: https://svn.erp5.org/repos/neo/trunk@1531 71dcc9de-d417-0410-9af5-da40c76e7ee4