Bug fix: Don't drop transaction information when locked.
If a client node disconnect from storage node right after send 'finishTransaction' to the primary master, then the storage clear it's related information and can not store them later in the database, when notifyUnlockInformation is received from the master.
Here the transaction object got a new attribute '_locked' set when the storage receive the 'lockInformation' from the master, meaning that the client is finish it's transaction.
If locked, the transaction is not cleared when the client disconnect from storage node.

git-svn-id: https://svn.erp5.org/repos/neo/trunk@1632 71dcc9de-d417-0410-9af5-da40c76e7ee4