Initial implementation of HandlerSwitcher for connections.
The main purpose is to apply and handler after all pending requests are
satisfied and allow process answer packets of current state out of order.
This fix a bug that appears when a storage tries to get a lock on an object that
is already held by a previous transaction. Is this case the store is delayed
causing the answer packet be sent out of order (packet sequence breakage), so
unexpected from client's point of view.

git-svn-id: https://svn.erp5.org/repos/neo/trunk@1874 71dcc9de-d417-0410-9af5-da40c76e7ee4