Fix the ability to run individual test files.
This was broken by r2371: it is good not to import it at file top-level,
but it must still be provided.
Also, fix 'No handler could be found for logger "NEO"' error messages,
which where hiding log messages, by using setupLog.
Also, manually set propagate to 0 so NEO messages are not output by root
logger too.

git-svn-id: https://svn.erp5.org/repos/neo/trunk@2376 71dcc9de-d417-0410-9af5-da40c76e7ee4