By default, don't drop locked transactions.
This change has an important effect on "abort" call in client handler: once
primary master node locked a transaction, a client must not be allowed to
abort it (it already asked to commit).
Also, remove parameter value in abortFor (it is the new default behaviour).
Also, document abort a bit more.

git-svn-id: https://svn.erp5.org/repos/neo/trunk@2417 71dcc9de-d417-0410-9af5-da40c76e7ee4