Ignore TID chronology when detecting & resolving conflicts.
If TID allocation and conflict resolution are serialised at database scope
(=cluster scope in NEO), the question doesn't arises: when we have a TID,
we are the latest TID until we call tpc_finish.
If they are not serialised, it means TID allocation will happen after
conflict resolution, as TID generation must be serialised anyway, and in
such case, the temporary TID we have when detecting conflicts has no
ordering signification, so it is meaningless to check their chronology.

git-svn-id: https://svn.erp5.org/repos/neo/trunk@2533 71dcc9de-d417-0410-9af5-da40c76e7ee4