Adding Neo buildout installation files
To allow Neo to be installed on a machine through buildout, we must define
buildout.cfg file. We added this fils to make it possible to set up a neo buildout on a machine
directly with repository sources. It is now possible to install neo on a computer
just by using buildout, linked to neo eggs.

Buildout docs

git-svn-id: https://svn.erp5.org/repos/neo/trunk@2659 71dcc9de-d417-0410-9af5-da40c76e7ee4