Restore an exception.
It is intended to raise even when running in optimised mode. The intent of
that change was to allow interpreter to detect the last elif is not
expected to be needed (although it is not known to make such deductions
currently). If only python had a deeper "enum" notion...
This (very) partially reverts f32e9120c3cdee8e91e11e35e30dc5747723b384.