storage: fix save or reset of 'backup_tid' config value
Because masters don't have persistent storage, the task to remember whether
the cluster is in backup mode or not is delegated to storages, via the presence
of a 'backup_tid' config value.

This fixes a bug that set 'backup_tid' after a simple replication.
If the cluster was restarted, it would have tried to switch to BACKINGUP state.