X neotest: Don't use bc
bc is marked as optional in Debian and I just hit a machine where it is
not installed. Python might be also not installed but we depend on
python in many more places so using it instead of bc reduces the
probability `neotest info-local` won't work on a fresh machine.