X neotest: Also benchmark NEO/py with logging disabled
NEO/py currently logs every packet internally in RAM buffer to flush on
SIGRTMIN if such a request comes. This might be adding overhead, so
Julien asked to also have numbers for NEO/py server with logging
disabled.

We already don't enable logging on NEO/py client since to enable it a
?logfile query needs to be passed to ZODB URL and we don't do so
anywhere.